kernel: sdup: removed sdup-common.h/c
