+++ linux/net/rina/Makefile
@@ -36,7 +36,6 @@ rina-personality-default-y :=					\
 	dtp-ps-default.o					\
 	dtcp-ps-default.o                       		\
 	pff-ps-default.o					\
-	sdup-ps-common.o					\
 	sdup-ps-default.o
 
 ifeq  ($(CONFIG_RINA_PERSONALITY_DEFAULT),m)
